process-mining/requirements.txt
google-cloud-documentai>=2.20.0
google-cloud-storage>=2.14.0
google-cloud-aiplatform>=1.38.0
vertexai>=0.0.1
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.9
redis>=5.0.1
